+++ src/main.js
@@ -4,7 +4,7 @@
 require('./form.js');
 require('./game.js');
 require('./reviews-block/reviews.js');
-var gallery = require('./const_gal.js');
+var gallery = require('./gallery.js');
 
 gallery.getPictures();
